+++ src/main/java/com/podium/service/RatingCategoryService.java
@@ -2,10 +2,14 @@ package com.podium.service;
 
 import com.podium.model.dto.request.RatingCategoryRequestDto;
 import com.podium.model.dto.response.RatingCategoryResponseDto;
+import com.podium.model.entity.City;
 import com.podium.model.entity.RatingCategory;
 import com.podium.repository.RatingCategoryRepository;
+import com.podium.service.exception.PodiumEntityAlreadyExistException;
+import com.podium.service.exception.PodiumEntityNotFoundException;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,46 +22,42 @@ public class RatingCategoryService {
         this.ratingCategoryRepository = ratingCategoryRepository;
     }
 
+    @Transactional
     public void addCategory(RatingCategoryRequestDto requestDto){
 
-        this.ratingCategoryRepository.save(
-                this.convertRequestDtoToEntity(requestDto));
+        if(this.ratingCategoryRepository.existsByCategory(requestDto.getCategory()))
+            throw new PodiumEntityAlreadyExistException("RatingDto Category");
 
+        this.ratingCategoryRepository.save(this.convertRequestDtoToEntity(requestDto));
     }
 
     public boolean existCategory(String category){
         return this.ratingCategoryRepository.existsByCategory(category);
     }
 
-    public Iterable<RatingCategoryResponseDto> findAll(){
+    public Iterable<RatingCategory> findAll(){
+        return this.ratingCategoryRepository.findAll();
+    }
 
-        List<RatingCategoryResponseDto> responseDtos = new ArrayList<>();
+    public RatingCategory findByCategory(String category){
 
-        this.ratingCategoryRepository
-                .findAll()
-                .forEach(x -> responseDtos
-                        .add(this.convertEntityToResponseDto(x))
-                );
-
-        return responseDtos;
+        return this.ratingCategoryRepository
+                .findByCategory(category).orElseThrow(() ->
+                new PodiumEntityNotFoundException("City"));
 
     }
 
-    public RatingCategoryResponseDto findByCategory(String category){
-        return this.convertEntityToResponseDto(
-                this.ratingCategoryRepository.findByCategory(category));
-    }
-
+    @Transactional
     public void deleteRatingCategory(String category){
+
+        if(!this.ratingCategoryRepository.existsByCategory(category))
+            throw new PodiumEntityNotFoundException("RatingDto Category");
+
         this.ratingCategoryRepository.deleteByCategory(category);
     }
 
     private RatingCategory convertRequestDtoToEntity(RatingCategoryRequestDto requestDto){
-        return  new RatingCategory(requestDto.getCategory());
-    }
-
-    private RatingCategoryResponseDto convertEntityToResponseDto(RatingCategory ratingCategory){
-        return new RatingCategoryResponseDto(ratingCategory.getCategory());
+        return new RatingCategory(requestDto.getCategory());
     }
 
 }
